packages/libs/test-utils/pyproject.toml
[tool.poetry]
name = "test-utils"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "test_utils"}]

[tool.poetry.dependencies]
python = "^3.11"
dagger-io = "^0.9.0"
pytest = "^7.4.2"

[project]
name = "test_utils"
version = "1.0.0" # ignore this, we use [tool.poetry]->version
classifiers = [
    "Framework :: Pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.entry-points.pytest11]
test_utils = "test_utils.fixtures"
